+++ src/fix/apply-auto-fixes.ts
@@ -1,6 +1,6 @@
 import { Rule } from 'eslint';
-import { notEmpty } from '../../util/type-check.js';
-import { FixContext } from '../index.js';
+import { notEmpty } from '../util/type-check.js';
+import { FixContext } from './index.js';
 
 // eslint-disable-next-line @typescript-eslint/ban-types
 export type FixToApplyAutoFixesArgs = {};
