+++ src/fix/apply-suggestions.test.ts
@@ -1,6 +1,6 @@
 import { basename } from 'node:path';
 import { describe, expect, test } from 'vitest';
-import { FixTester } from '../../test-util/fix-tester.js';
+import { FixTester } from '../test-util/fix-tester.js';
 import { createFixToApplySuggestions, FixToApplySuggestionsArgs } from './apply-suggestions.js';
 
 const tester = new FixTester<FixToApplySuggestionsArgs>(
