+++ src/fix/apply-suggestions.ts
@@ -1,5 +1,5 @@
 import { Linter, Rule } from 'eslint';
-import { FixContext } from '../index.js';
+import { FixContext } from './index.js';
 
 export type SuggestionFilter = (
   suggestions: Linter.LintSuggestion[],
