+++ src/fix/disable-per-file.ts
@@ -1,7 +1,7 @@
 import { Rule, SourceCode } from 'eslint';
 import { DescriptionPosition } from 'src/cli/prompt.js';
-import { mergeFixes } from '../../eslint/report-translator.js';
-import { unique } from '../../util/array.js';
+import { mergeFixes } from '../eslint/report-translator.js';
+import { unique } from '../util/array.js';
 import {
   DisableComment,
   findShebang,
@@ -11,9 +11,9 @@ import {
   mergeDescription,
   mergeRuleIds,
   updateDisableComment,
-} from '../../util/eslint.js';
-import { notEmpty } from '../../util/type-check.js';
-import { FixContext } from '../index.js';
+} from '../util/eslint.js';
+import { notEmpty } from '../util/type-check.js';
+import { FixContext } from './index.js';
 
 export type FixToDisablePerFileArgs = {
   description?: string | undefined;
