+++ src/index.ts
@@ -2,4 +2,4 @@ export { run, type Options } from './cli/run.js';
 export { Core } from './core.js';
 export { type Config, type ESLintOptions, configDefaults } from './config.js';
 export { takeRuleStatistics, type RuleStatistic } from './formatter/index.js';
-export { type FixableMaker, type SuggestionFilter, type FixContext } from './fixer/index.js';
+export { type FixableMaker, type SuggestionFilter, type FixContext } from './fix/index.js';
